+++ src/main.js
@@ -11,23 +11,43 @@
  */
 
 
+/**
+ * @typedef {Object} Batch
+ * @property {string} name - Batch name
+ * @property {function} source - Data source
+ * @property {function} job - Job to run on each data item
+ * 
+ */
+
 
 function batchRunner () {
 
     const store = new Map();
 
 
-
+    /**
+     * @function define
+     * @description Define a batch
+     * @param {Batch} batch - Batch definition
+     * @returns {boolean} - true if batch was defined, false otherwise
+     */
     function define ( batch ) {
                 const { name, source, job } = batch;
                 if ( typeof name   !== 'string'   )   return false
                 if ( typeof source !== 'function' )   return false
                 if ( typeof job    !== 'function' )   return false
                 store.set( name, batch )
+                return true
         } // define func.
 
 
-
+    /**
+     * @function run
+     * @description Run a batch over each of data items
+     * @param {string} name - Batch name
+     * @param {...*} [args] - Extra arguments passed to the job
+     * @returns {Array} - Array of results
+     */
     function run ( name, ...args ) {
 
         if ( typeof name !== 'string') {   // When we want to register a new batch and run it with the same call
@@ -37,7 +57,7 @@ function batchRunner () {
             }
 
         const record = store.get( name );
-        if ( record == null ) return false
+        if ( record == null )   return []
 
         const { source, job } = record;
         let data = source ();