+++ Changelog.md
@@ -1,5 +1,12 @@
 # Release History
 
 
+### 1.0.2 ( 2023-11-21)
+- [x] Type defintions;
+- [x] Fix: Method run() should always return an array;
+
+
+
 ## 1.0.0 (2023-11-19)
 - [x] Initial release
+- [] Bug: Method run() should always return an array. When record not found, returns false;